refactor(client): extract p2p logic to state::p2p
