+++ client/src/actors.rs
@@ -9,7 +9,7 @@ mod snapshot;
 
 #[cfg(feature = "p2p")]
 pub use self::{
-    p2p::{messages as network_messages, NetworkActor, NetworkConfig},
+    p2p::messages as network_messages,
     registry::p2p_messages::{GetNetwork, InsertNetwork, RemoveNetwork},
     secure::p2p_messages as client_p2p_messages,
 };
