+++ client/src/state.rs
@@ -2,5 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod key_store;
+#[cfg(feature = "p2p")]
+pub mod p2p;
 pub mod secure;
 pub mod snapshot;
